backend/requirements.txt
flask==3.0.3
werkzeug==3.0.3
httpx[http2]==0.27.0
pydantic==2.8.2
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
nltk==3.9.1
pypdf==4.3.1
python-docx==1.1.2
